# openapi.yaml
openapi: 3.0.0
info:
  title: reciclaapp API
  description: API REST completa con autenticación JWT y CRUD de entidades
  version: 1.0.0
servers:
  - url: http://localhost:3016/api
    description: Servidor de desarrollo
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EntityBase:
      type: object
      properties:
        _id:
          type: string
        nombre:
          type: string
        descripcion:
          type: string
        imagen:
          type: string
        imagenThumbnail:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
...
paths:
...
  /material-reciclado:
    get:
      tags:
        - MaterialReciclado
      summary: Listar todos los MaterialReciclados
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de MaterialReciclados
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - MaterialReciclado
      summary: Crear MaterialReciclado
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: MaterialReciclado creado
  /material-reciclado/{id}:
    get:
      tags:
        - MaterialReciclado
      summary: Obtener MaterialReciclado por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: MaterialReciclado encontrado
        '404':
          description: MaterialReciclado no encontrado
    patch:
      tags:
        - MaterialReciclado
      summary: Actualizar MaterialReciclado
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: MaterialReciclado actualizado
    delete:
      tags:
        - MaterialReciclado
      summary: Eliminar MaterialReciclado
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: MaterialReciclado eliminado
  /material-reciclado/{id}/upload-image:
    post:
      tags:
        - MaterialReciclado
      summary: Subir imagen a MaterialReciclado
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida
  /punto-recoleccion:
    get:
      tags:
        - PuntoRecoleccion
      summary: Listar todos los PuntoRecoleccions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de PuntoRecoleccions
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - PuntoRecoleccion
      summary: Crear PuntoRecoleccion
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: PuntoRecoleccion creado
  /punto-recoleccion/{id}:
    get:
      tags:
        - PuntoRecoleccion
      summary: Obtener PuntoRecoleccion por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: PuntoRecoleccion encontrado
        '404':
          description: PuntoRecoleccion no encontrado
    patch:
      tags:
        - PuntoRecoleccion
      summary: Actualizar PuntoRecoleccion
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: PuntoRecoleccion actualizado
    delete:
      tags:
        - PuntoRecoleccion
      summary: Eliminar PuntoRecoleccion
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: PuntoRecoleccion eliminado
  /punto-recoleccion/{id}/upload-image:
    post:
      tags:
        - PuntoRecoleccion
      summary: Subir imagen a PuntoRecoleccion
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida
...